feat: 단일 객체 조회 메서드 queryForObject 구현

+++ jdbc/src/test/java/com/interface21/jdbc/core/RowMapperTest.java
@@ -31,6 +31,6 @@ class TestUserRowMapper implements RowMapper<TestUser> {
 
     @Override
     public TestUser mapRow(ResultSet resultSet, int rowNumber) throws SQLException {
-        return new TestUser(resultSet.getInt("id"), resultSet.getString("name"));
+        return new TestUser(resultSet.getLong("id"), resultSet.getString("name"));
     }
 }
